# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Edna Bar Book Printing API
  description: |
    A comprehensive RESTful API for managing book printing operations.
    
    This API provides complete functionality for:
    - Managing books, authors, and publishers
    - Handling printing companies (private and contracted)
    - Creating and tracking printing contracts
    - Scheduling and monitoring printing jobs
    - Real-time delivery tracking
    
    Built with Go following clean architecture principles.
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/edna-bar/api
    email: support@edna-bar.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://edna-bar.com/terms
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.edna-bar.com/api
    description: Production server
...
paths:
...
  /editoras:
    get:
      tags:
        - publishers
      summary: List all publishers
      description: Retrieve a list of all publishers in the system
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publisher'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - publishers
      summary: Create a new publisher
      description: Add a new publisher to the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePublisherRequest'
      responses:
        '201':
          description: Publisher created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publisher'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /editoras/{id}:
    get:
      tags:
        - publishers
      summary: Get publisher by ID
      description: Retrieve a specific publisher by ID
      parameters:
        - name: id
          in: path
          required: true
          description: Publisher ID
          schema:
            type: integer
      responses:
        '200':
          description: Publisher found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publisher'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - publishers
      summary: Update publisher
      description: Update an existing publisher
      parameters:
        - name: id
          in: path
          required: true
          description: Publisher ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePublisherRequest'
      responses:
        '200':
          description: Publisher updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publisher'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - publishers
      summary: Delete publisher
      description: Remove a publisher from the system
      parameters:
        - name: id
          in: path
          required: true
          description: Publisher ID
          schema:
            type: integer
      responses:
        '204':
          description: Publisher deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Printing Companies endpoints
  /graficas:
    get:
      tags:
        - printing-companies
      summary: List all printing companies
      description: Retrieve a list of all printing companies
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PrintingCompany'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - printing-companies
      summary: Create a new printing company
      description: Add a new printing company to the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePrintingCompanyRequest'
      responses:
        '201':
          description: Printing company created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrintingCompany'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /graficas/{id}:
    get:
      tags:
        - printing-companies
      summary: Get printing company by ID
      description: Retrieve a specific printing company by ID
      parameters:
        - name: id
          in: path
          required: true
          description: Printing company ID
          schema:
            type: integer
      responses:
        '200':
          description: Printing company found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrintingCompany'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Contracts endpoints
...
components:
  schemas:
...
    Publisher:
      type: object
      properties:
        id:
          type: integer
          description: Publisher ID
          example: 1
        nome:
          type: string
          description: Publisher name
          example: "Penguin Random House Brasil"
        endereco:
          type: string
          description: Publisher address
          example: "São Paulo, SP, Brasil"
      required:
        - id
        - nome
        - endereco
    CreatePublisherRequest:
      type: object
      properties:
        nome:
          type: string
          description: Publisher name
          maxLength: 255
          example: "Penguin Random House Brasil"
        endereco:
          type: string
          description: Publisher address
          maxLength: 255
          example: "São Paulo, SP, Brasil"
      required:
        - nome
        - endereco
    UpdatePublisherRequest:
      type: object
      properties:
        nome:
          type: string
          description: Publisher name
          maxLength: 255
          example: "Penguin Random House Brasil"
        endereco:
          type: string
          description: Publisher address
          maxLength: 255
          example: "São Paulo, SP, Brasil"
    PrintingCompany:
      type: object
      properties:
        id:
          type: integer
          description: Printing company ID
          example: 1
        nome:
          type: string
          description: Printing company name
          example: "PrintTech Solutions"
        type:
          type: string
          enum: [particular, contratada]
          description: Type of printing company
          example: "contratada"
        endereco:
          type: string
          description: Address (for contracted companies)
          example: "Distrito Industrial, São Paulo, SP"
      required:
        - id
        - nome
        - type
    CreatePrintingCompanyRequest:
      type: object
      properties:
        nome:
          type: string
          description: Printing company name
          maxLength: 255
          example: "PrintTech Solutions"
        type:
          type: string
          enum: [particular, contratada]
          description: Type of printing company
          example: "contratada"
        endereco:
          type: string
          description: Address (required for contracted companies)
          maxLength: 255
          example: "Distrito Industrial, São Paulo, SP"
      required:
        - nome
        - type
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Resource not found"
        code:
          type: string
          description: Error code
          example: "NOT_FOUND"
        details:
          type: string
          description: Additional error details
          example: "Book with ISBN 978-0123456789 was not found"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-15T10:30:00Z"
      required:
        - error
        - code
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid input data"
            code: "BAD_REQUEST"
            details: "ISBN must be a valid format"
            timestamp: "2024-01-15T10:30:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Resource not found"
            code: "NOT_FOUND"
            details: "Book with ISBN 978-0123456789 was not found"
            timestamp: "2024-01-15T10:30:00Z"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            code: "INTERNAL_ERROR"
            details: "An unexpected error occurred"
            timestamp: "2024-01-15T10:30:00Z"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

# Global security (can be overridden per endpoint)
# security:
#   - bearerAuth: []